requirements.txt
datasets==3.6.0
gradio==5.33.0
langchain==0.3.25
langchain_core==0.3.64
langchain_openai==0.3.21
python-dotenv==1.1.0
soundfile==0.13.1
torch==2.6.0
transformers==4.52.4
openai
sentencepiece
pillow